Parametrize and merge resolve_harp tests into test_ephys_utils_unit.py
